imdb_api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.4.1
    # via django
astroid==2.9.0
    # via pylint
attrs==21.2.0
    # via pytest
black==21.11b1
    # via -r requirements.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via black
colorama==0.4.4
    # via crayons
configparser==5.2.0
    # via webdriver-manager
crayons==0.4.0
    # via webdriver-manager
django-cors-headers==3.10.0
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via
    #   -r requirements.in
    #   djangorestframework-filters
django==3.2.9
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework-filters==1.0.0.dev0
    # via -r requirements.in
djangorestframework-simplejwt==5.0.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-filters
    #   djangorestframework-simplejwt
factory-boy==3.2.1
    # via -r requirements.in
faker==9.9.0
    # via factory-boy
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via pylint
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via pylint
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pep8==1.7.1
    # via -r requirements.in
pillow==8.4.0
    # via -r requirements.in
platformdirs==2.4.0
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
psycopg2==2.9.2
    # via -r requirements.in
py==1.11.0
    # via pytest
pyjwt==2.3.0
    # via djangorestframework-simplejwt
pylint==2.12.1
    # via -r requirements.in
pyparsing==3.0.6
    # via packaging
pytest-django==4.5.0
    # via -r requirements.in
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-django
python-dateutil==2.8.2
    # via faker
pytz==2021.3
    # via django
regex==2021.11.10
    # via black
requests==2.26.0
    # via webdriver-manager
six==1.16.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlparse==0.4.2
    # via django
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   pylint
    #   pytest
tomli==1.2.2
    # via black
typing-extensions==4.0.1
    # via
    #   astroid
    #   black
    #   pylint
urllib3==1.26.7
    # via requests
webdriver-manager==3.5.2
    # via -r requirements.in
wrapt==1.13.3
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
